Don't Override AppAlreadyRunning when it is provided via the dataFields
Summary:

The API requires that AppAlreadyRunning be provided as a parameter when trackingPushOpen. It also allows for a DataFields parameter. These are merged together. Before, the provided AppAlreadyRunning overwrote the value in the DataFields. I don't think that is correct, and instead the DataFields' key takes priority.

Reviewers:
Brophy

Test Plan:

Revert Plan:
git revert